Added Save mechanic, Fixed Lava path damage
I made a checkpoint save mechanic, so that the game saves afteer the cut scene, so u can continue from there if you die.
I fixed the lava paths, so that they now do damage, when in lava state and dont do damage in path state.
I also opened the paths, so that now u can walk of the sides of the path, and get damaged when you try to do so.